+++ rumqttc/src/v5/state.rs
@@ -65,7 +65,7 @@ pub enum StateError {
     #[error("Connection failed with reason '{reason:?}' ")]
     ConnFail { reason: ConnectReturnCode },
     #[error("Connection closed by peer abruptly")]
-    ConnectionAborted
+    ConnectionAborted,
 }
 
 impl From<mqttbytes::Error> for StateError {
@@ -138,7 +138,7 @@ impl MqttState {
             // index 0 is wasted as 0 is not a valid packet id
             outgoing_pub: vec![None; max_inflight as usize + 1],
             outgoing_rel: vec![None; max_inflight as usize + 1],
-            incoming_pub: vec![None; std::u16::MAX as usize + 1],
+            incoming_pub: vec![None; u16::MAX as usize + 1],
             collision: None,
             // TODO: Optimize these sizes later
             events: VecDeque::with_capacity(100),
@@ -217,6 +217,8 @@ impl MqttState {
         &mut self,
         mut packet: Incoming,
     ) -> Result<Option<Packet>, StateError> {
+        self.events.push_back(Event::Incoming(packet.to_owned()));
+
         let outgoing = match &mut packet {
             Incoming::PingResp(_) => self.handle_incoming_pingresp()?,
             Incoming::Publish(publish) => self.handle_incoming_publish(publish)?,
@@ -234,7 +236,6 @@ impl MqttState {
             }
         };
 
-        self.events.push_back(Event::Incoming(packet));
         self.last_incoming = Instant::now();
         Ok(outgoing)
     }
@@ -331,7 +332,7 @@ impl MqttState {
             }
         } else if let Some(alias) = topic_alias {
             if let Some(topic) = self.topic_alises.get(&alias) {
-                publish.topic = topic.to_owned();
+                topic.clone_into(&mut publish.topic);
             } else {
                 self.handle_protocol_error()?;
             };